Replace NextTickList implementation
Vanilla uses a set and a sorted map in order to prioritize the order
of blocks being ticked in the world and provide fast lookup, the problem
here is that vanilla will, in some cases, attempt to reschedule a block
for work, and while by equals will tolerate this, the comparator used in
NextTickListEntry will never equal 0 for such tasks, meaning that any collection
which attempts to optimize searching its bins using the caparator will fail to
find the existing value, causing desyncronization between these two collections.

We attempt to solve this by reusing the older tick list entry when attempting
to remove any older values. This patch is intended to be as lighetweight
as possible, while still supporting anything that may be attempting to access
this list using reflection.